Simplify test and move to test_converter.py file
